Remove old PIN code (#223)
