+++ app/src/main/AndroidManifest.xml
@@ -61,7 +61,7 @@
             android:name=".modules.settings.AboutSettingsActivity"
             android:screenOrientation="portrait"/>
         <activity
-            android:name=".modules.newpin.NewPinActivity"
+            android:name=".modules.pin.PinActivity"
             android:launchMode="singleTask"
             android:screenOrientation="portrait"/>
