+++ app/src/main/java/bitcoin/wallet/modules/pin/ManagePinPresenter.kt
@@ -1,4 +1,4 @@
-package bitcoin.wallet.modules.newpin
+package bitcoin.wallet.modules.pin
 
 import bitcoin.wallet.R
