+++ app/src/main/java/bitcoin/wallet/modules/pin/PinActivity.kt
@@ -6,10 +6,14 @@ import android.arch.lifecycle.ViewModelProviders
 import android.content.Context
 import android.content.Intent
 import android.hardware.fingerprint.FingerprintManager
-import android.os.*
+import android.os.Bundle
+import android.os.Handler
 import android.security.keystore.KeyPermanentlyInvalidatedException
 import android.security.keystore.UserNotAuthenticatedException
+import android.support.constraint.ConstraintLayout
 import android.support.v7.widget.GridLayoutManager
+import android.support.v7.widget.LinearLayoutManager
+import android.support.v7.widget.PagerSnapHelper
 import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
@@ -24,152 +28,48 @@ import bitcoin.wallet.core.App
 import bitcoin.wallet.core.security.EncryptionManager
 import bitcoin.wallet.core.security.FingerprintAuthenticationDialogFragment
 import bitcoin.wallet.modules.main.MainModule
+import bitcoin.wallet.ui.extensions.SmoothLinearLayoutManager
 import bitcoin.wallet.viewHelpers.HudHelper
-import kotlinx.android.synthetic.main.activity_pin.*
+import kotlinx.android.synthetic.main.activity_new_pin.*
 import kotlinx.android.synthetic.main.custom_tall_toolbar.*
 import java.security.UnrecoverableKeyException
-
-
-
 
 class PinActivity : BaseActivity(), NumPadItemsAdapter.Listener, FingerprintAuthenticationDialogFragment.Callback {
 
     private lateinit var viewModel: PinViewModel
-
-    private lateinit var imgPinMask1: ImageView
-    private lateinit var imgPinMask2: ImageView
-    private lateinit var imgPinMask3: ImageView
-    private lateinit var imgPinMask4: ImageView
-    private lateinit var imgPinMask5: ImageView
-    private lateinit var imgPinMask6: ImageView
-
-    private var isFingerprintEnabled = false
+    private lateinit var layoutManager: LinearLayoutManager
+    private lateinit var pinPagesAdapter: PinPagesAdapter
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
         window.setFlags(WindowManager.LayoutParams.FLAG_SECURE, WindowManager.LayoutParams.FLAG_SECURE)
 
-        setContentView(R.layout.activity_pin)
+        setContentView(R.layout.activity_new_pin)
 
         setSupportActionBar(toolbar)
-        supportActionBar?.title = ""
-
-        val interactionType = intent.getSerializableExtra(keyInteractionType) as PinInteractionType
-        val enteredPin = intent.getStringExtra(keyEnteredPin)
-
+        backButton.visibility = View.GONE
         backButton.setOnClickListener { viewModel.delegate.onBackPressed() }
 
+        val interactionType = intent.getSerializableExtra(keyInteractionType) as PinInteractionType
+
+        pinPagesAdapter = PinPagesAdapter()
+        layoutManager = SmoothLinearLayoutManager(this, LinearLayoutManager.HORIZONTAL, false)
+        pinPagesRecyclerView.layoutManager = layoutManager
+        val snapHelper = PagerSnapHelper()
+        snapHelper.attachToRecyclerView(pinPagesRecyclerView)
+        pinPagesRecyclerView.adapter = pinPagesAdapter
+
+        pinPagesRecyclerView.setOnTouchListener { _, _ ->
+            //disable recyclerview scroll
+            true
+        }
+
+
         viewModel = ViewModelProviders.of(this).get(PinViewModel::class.java)
-        viewModel.init(interactionType, enteredPin)
+        viewModel.init(interactionType)
 
-        viewModel.title.observe(this, Observer { title ->
-            title?.let {
-                toolbarTitle.text = getString(it)
-            }
-        })
-
-        viewModel.description.observe(this, Observer { description ->
-            description?.let { txtPromptPin.setText(it) }
-        })
-
-        viewModel.highlightPinMask.observe(this, Observer { length ->
-            length?.let { updatePinCircles(it) }
-        })
-
-        viewModel.showErrorInDialog.observe(this, Observer { error ->
-            error?.let {
-                HudHelper.showErrorMessage(error, this)
-            }
-        })
-
-        viewModel.showErrorMessage.observe(this, Observer { error ->
-            error?.let {
-                errorMessage.setText(error)
-            }
-        })
-
-        viewModel.goToPinConfirmation.observe(this, Observer { pin ->
-            pin?.let {
-                PinModule.startForSetPinConfirm(this, pin)
-            }
-        })
-
-        viewModel.goToSetPin.observe(this, Observer {
-            PinModule.startForSetPin(this)
-        })
-
-        viewModel.showSuccess.observe(this, Observer {
-            HudHelper.showSuccessMessage(activity = this)
-            MainModule.start(this)
-            finish()
-        })
-
-        viewModel.hideToolbar.observe(this, Observer {
-            supportActionBar?.hide()
-        })
-
-        viewModel.unlockWallet.observe(this, Observer {
-            unlockWallet()
-        })
-
-        viewModel.clearPinMaskWithDelay.observe(this, Observer {
-            Handler().postDelayed({
-                updatePinCircles(0)
-            }, 300)
-        })
-
-        viewModel.showFingerprintDialog.observe(this, Observer {
-            showFingerprintDialog()
-        })
-
-        viewModel.minimizeApp.observe(this, Observer {
-            moveTaskToBack(true)
-        })
-
-        viewModel.goBack.observe(this, Observer {
-            super.onBackPressed()
-        })
-
-        viewModel.goToPinEdit.observe(this, Observer {
-            PinModule.startForEditPin(this)
-            finish()
-        })
-
-        viewModel.hideBackButton.observe(this, Observer {
-            backButton.visibility = View.GONE
-        })
-
-        viewModel.keyStoreSafeExecute.observe(this, Observer { triple ->
-            triple?.let { (action, onSuccess, onFailure) ->
-                safeExecuteWithKeystore(action, onSuccess, onFailure)
-            }
-        })
-
-        viewModel.onWrongPin.observe(this, Observer {
-            vibrate()
-            shakePinCircles()
-        })
-
-        viewModel.blockScreen.observe(this, Observer {
-            pinUnlock.visibility = View.GONE
-            pinUnlockBlocked.visibility = View.VISIBLE
-        })
-
-        viewModel.unblockScreen.observe(this, Observer {
-            pinUnlock.visibility = View.VISIBLE
-            pinUnlockBlocked.visibility = View.GONE
-        })
-
-
-        imgPinMask1 = findViewById(R.id.imgPinMaskOne)
-        imgPinMask2 = findViewById(R.id.imgPinMaskTwo)
-        imgPinMask3 = findViewById(R.id.imgPinMaskThree)
-        imgPinMask4 = findViewById(R.id.imgPinMaskFour)
-        imgPinMask5 = findViewById(R.id.imgPinMaskFive)
-        imgPinMask6 = findViewById(R.id.imgPinMaskSix)
-
-        numPadItems.adapter = NumPadItemsAdapter(listOf(
+        val numpadAdapter = NumPadItemsAdapter(listOf(
                 NumPadItem(NumPadItemType.NUMBER, 1, ""),
                 NumPadItem(NumPadItemType.NUMBER, 2, "abc"),
                 NumPadItem(NumPadItemType.NUMBER, 3, "def"),
@@ -184,15 +84,121 @@ class PinActivity : BaseActivity(), NumPadItemsAdapter.Listener, FingerprintAuth
                 NumPadItem(NumPadItemType.DELETE, 0, "DEL")
         ), this)
 
+        numPadItems.adapter = numpadAdapter
         numPadItems.layoutManager = GridLayoutManager(this, 3)
+
+        viewModel.titleLiveDate.observe(this, Observer { title ->
+            title?.let { toolbarTitle.text = getString(it) }
+        })
+
+        viewModel.hideToolbar.observe(this, Observer {
+            supportActionBar?.hide()
+        })
+
+        viewModel.showBackButton.observe(this, Observer {
+            backButton.visibility = View.VISIBLE
+        })
+
+        viewModel.titleLiveDate.observe(this, Observer { title ->
+            title?.let { supportActionBar?.title = getString(it) }
+        })
+
+        viewModel.addPagesEvent.observe(this, Observer { pinPages ->
+            pinPages?.let {
+                pinPagesAdapter.pinPages.addAll(it)
+                pinPagesAdapter.notifyDataSetChanged()
+            }
+        })
+
+        viewModel.showPageAtIndex.observe(this, Observer { index ->
+            index?.let {
+                Handler().postDelayed({
+                    pinPagesAdapter.setEnteredPinLength(layoutManager.findFirstVisibleItemPosition(), 0)
+                    pinPagesRecyclerView.smoothScrollToPosition(it)
+                }, 300)
+            }
+        })
+
+        viewModel.showErrorForPage.observe(this, Observer { errorForPage ->
+            errorForPage?.let { (error, pageIndex) ->
+                pinPagesAdapter.setErrorForPage(pageIndex, error)
+            }
+        })
+
+        viewModel.showError.observe(this, Observer { error ->
+            error?.let {
+                HudHelper.showErrorMessage(it, this)
+            }
+        })
+
+        viewModel.navigateToMainLiveEvent.observe(this, Observer {
+            Handler().postDelayed({
+                HudHelper.showSuccessMessage(activity = this)
+                MainModule.start(this)
+                finish()
+            }, 300)
+        })
+
+        viewModel.keyStoreSafeExecute.observe(this, Observer { triple ->
+            triple?.let {
+                val (action, onSuccess, onFailure) = it
+                safeExecuteWithKeystore(action, onSuccess, onFailure)
+            }
+        })
+
+        viewModel.fillPinCircles.observe(this, Observer { pair ->
+            pair?.let { (length, pageIndex) ->
+                pinPagesAdapter.setEnteredPinLength(pageIndex, length)
+            }
+        })
+
+        viewModel.dismissLiveEvent.observe(this, Observer {
+            finish()
+        })
+
+        viewModel.showSuccess.observe(this, Observer {
+            HudHelper.showSuccessMessage(activity = this)
+        })
+
+        viewModel.showFingerprintInputLiveEvent.observe(this, Observer {
+            showFingerprintDialog()
+            numpadAdapter.showFingerPrintButton = true
+        })
+
+        viewModel.resetCirclesWithShakeAndDelayForPage.observe(this, Observer { pageIndex ->
+            pageIndex?.let {
+                pinPagesAdapter.shakePageIndex = it
+                pinPagesAdapter.notifyDataSetChanged()
+                Handler().postDelayed({
+                    pinPagesAdapter.shakePageIndex = null
+                    pinPagesAdapter.setEnteredPinLength(pageIndex, 0)
+                    viewModel.delegate.resetPin()
+                }, 300)
+            }
+        })
+
     }
 
     override fun onBackPressed() {
         viewModel.delegate.onBackPressed()
     }
 
+    override fun onItemClick(item: NumPadItem) {
+        when (item.type) {
+            NumPadItemType.NUMBER -> {
+                viewModel.delegate.onEnter(item.number.toString(), layoutManager.findFirstVisibleItemPosition())
+            }
+            NumPadItemType.DELETE -> {
+                viewModel.delegate.onDelete(layoutManager.findFirstVisibleItemPosition())
+            }
+            NumPadItemType.FINGER -> {
+                viewModel.delegate.showBiometricUnlock()
+            }
+        }
+    }
+
     override fun onFingerprintAuthSucceed(withFingerprint: Boolean, crypto: FingerprintManager.CryptoObject?) {
-        unlockWallet()
+        viewModel.delegate.onBiometricUnlock()
     }
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
@@ -205,25 +211,6 @@ class PinActivity : BaseActivity(), NumPadItemsAdapter.Listener, FingerprintAuth
         super.onActivityResult(requestCode, resultCode, data)
     }
 
-    override fun onItemClick(item: NumPadItem) {
-        when (item.type) {
-            NumPadItemType.NUMBER -> {
-                errorMessage.text = ""
-                viewModel.delegate.onEnterDigit(item.number)
-            }
-            NumPadItemType.DELETE -> {
-                viewModel.delegate.onClickDelete()
-            }
-            NumPadItemType.FINGER -> {
-                if (isFingerprintEnabled) {
-                    showFingerprintDialog()
-                }
-            }
-        }
-    }
-
-    override fun isFingerPrintEnabled() = isFingerprintEnabled
-
     private fun showFingerprintDialog() {
         if (App.systemInfoManager.touchSensorCanBeUsed()) {
             try {
@@ -234,10 +221,7 @@ class PinActivity : BaseActivity(), NumPadItemsAdapter.Listener, FingerprintAuth
                 fragment.isCancelable = true
                 fragment.show(fragmentManager, "fingerprint_dialog")
 
-                isFingerprintEnabled = true
-                numPadItems.adapter?.notifyDataSetChanged()
-
-            } catch (e: Exception) {
+            } catch (e: Throwable) {
                 when (e) {
                     is UserNotAuthenticatedException -> EncryptionManager.showAuthenticationScreen(this, AUTHENTICATE_TO_FINGERPRINT)
                     is KeyPermanentlyInvalidatedException,
@@ -247,28 +231,82 @@ class PinActivity : BaseActivity(), NumPadItemsAdapter.Listener, FingerprintAuth
         }
     }
 
-    private fun shakePinCircles() {
-        val shake = AnimationUtils.loadAnimation(this, R.anim.shake_pin_circles)
-        pinCirclesWrapper.startAnimation(shake)
+    companion object {
+        const val AUTHENTICATE_TO_FINGERPRINT = 1
+
+        private const val keyInteractionType = "interaction_type"
+
+        fun start(context: Context, interactionType: PinInteractionType) {
+            val intent = Intent(context, PinActivity::class.java)
+            intent.putExtra(keyInteractionType, interactionType)
+            context.startActivity(intent)
+        }
+
+        fun startForUnlock() {
+            val intent = Intent(App.instance, PinActivity::class.java)
+            intent.putExtra(keyInteractionType, PinInteractionType.UNLOCK)
+            intent.flags = Intent.FLAG_ACTIVITY_EXCLUDE_FROM_RECENTS and Intent.FLAG_ACTIVITY_NEW_TASK and Intent.FLAG_ACTIVITY_CLEAR_TOP
+            App.instance.startActivity(intent)
+        }
+    }
+}
+
+enum class NumPadItemType {
+    NUMBER, DELETE, FINGER
+}
+
+//PinPage part
+class PinPage(val description: Int, var enteredDidgitsLength: Int = 0, var error: Int? = null)
+
+class PinPagesAdapter : RecyclerView.Adapter<RecyclerView.ViewHolder>() {
+
+    var pinPages = mutableListOf<PinPage>()
+    var shakePageIndex: Int? = null
+
+    fun setErrorForPage(pageIndex: Int, error: Int?) {
+        pinPages[pageIndex].error = error
+        notifyDataSetChanged()
     }
 
-    private fun vibrate() {
-        val v = getSystemService(Context.VIBRATOR_SERVICE) as Vibrator
-        val vibrateTime = 300L
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            v.vibrate(VibrationEffect.createOneShot(vibrateTime, VibrationEffect.DEFAULT_AMPLITUDE))
-        } else {//deprecated in API 26
-            v.vibrate(vibrateTime)
+    fun setEnteredPinLength(pageIndex: Int, enteredLength: Int) {
+        pinPages[pageIndex].enteredDidgitsLength = enteredLength
+        notifyDataSetChanged()
+    }
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
+        return PinPageViewHolder(LayoutInflater.from(parent.context).inflate(R.layout.view_pin_page, parent, false))
+    }
+
+    override fun getItemCount() = pinPages.count()
+
+    override fun onBindViewHolder(holder: RecyclerView.ViewHolder, position: Int) {
+        if (holder is PinPageViewHolder) {
+            holder.bind(pinPages[position], shakePageIndex == position)//, { listener.onItemClick(numPadItems[position]) }, listener.isFingerPrintEnabled())
         }
     }
 
-    private fun unlockWallet() {
-        finish()
-    }
+}
 
-    override fun onSupportNavigateUp(): Boolean {
-        viewModel.delegate.onBackPressed()
-        return true
+class PinPageViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
+    private var txtDesc: TextView = itemView.findViewById(R.id.txtDescription)
+    private var txtError: TextView = itemView.findViewById(R.id.errorMessage)
+    private var pinCirclesWrapper = itemView.findViewById<ConstraintLayout>(R.id.pinCirclesWrapper)
+
+    private var imgPinMask1: ImageView = itemView.findViewById(R.id.imgPinMaskOne)
+    private var imgPinMask2: ImageView = itemView.findViewById(R.id.imgPinMaskTwo)
+    private var imgPinMask3: ImageView = itemView.findViewById(R.id.imgPinMaskThree)
+    private var imgPinMask4: ImageView = itemView.findViewById(R.id.imgPinMaskFour)
+    private var imgPinMask5: ImageView = itemView.findViewById(R.id.imgPinMaskFive)
+    private var imgPinMask6: ImageView = itemView.findViewById(R.id.imgPinMaskSix)
+
+    fun bind(pinPage: PinPage, shake: Boolean) {
+        txtDesc.text = itemView.resources.getString(pinPage.description)
+        updatePinCircles(pinPage.enteredDidgitsLength)
+        txtError.text = pinPage.error?.let { itemView.resources.getString(it) } ?: ""
+        if (shake) {
+            val shakeAnim = AnimationUtils.loadAnimation(itemView.context, R.anim.shake_pin_circles)
+            pinCirclesWrapper.startAnimation(shakeAnim)
+        }
     }
 
     private fun updatePinCircles(length: Int) {
@@ -282,35 +320,16 @@ class PinActivity : BaseActivity(), NumPadItemsAdapter.Listener, FingerprintAuth
         imgPinMask5.setImageResource(if (length > 4) filledCircle else emptyCircle)
         imgPinMask6.setImageResource(if (length > 5) filledCircle else emptyCircle)
     }
-
-    companion object {
-
-        const val AUTHENTICATE_TO_FINGERPRINT = 1
-
-        private const val keyInteractionType = "interaction_type"
-        private const val keyEnteredPin = "entered_pin"
-
-        fun start(context: Context, interactionType: PinInteractionType, enteredPin: String = "") {
-//            val intent = Intent(context, PinActivity::class.java)
-//            intent.putExtra(keyInteractionType, interactionType)
-//            intent.putExtra(keyEnteredPin, enteredPin)
-//            context.startActivity(intent)
-        }
-    }
-
 }
 
-enum class NumPadItemType {
-    NUMBER, DELETE, FINGER
-}
-
+//NumPad part
 data class NumPadItem(val type: NumPadItemType, val number: Int, val letters: String)
 
-class NumPadItemsAdapter(private val numPadItems: List<NumPadItem>, private val listener: Listener) : RecyclerView.Adapter<RecyclerView.ViewHolder>() {
+class NumPadItemsAdapter(private val numPadItems: List<NumPadItem>, private val listener: Listener)
+    : RecyclerView.Adapter<RecyclerView.ViewHolder>() {
 
     interface Listener {
         fun onItemClick(item: NumPadItem)
-        fun isFingerPrintEnabled(): Boolean
     }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
@@ -319,9 +338,15 @@ class NumPadItemsAdapter(private val numPadItems: List<NumPadItem>, private val 
 
     override fun getItemCount() = numPadItems.count()
 
+    var showFingerPrintButton = false
+        set(value) {
+            field = value
+            notifyDataSetChanged()
+        }
+
     override fun onBindViewHolder(holder: RecyclerView.ViewHolder, position: Int) {
         if (holder is NumPadItemViewHolder) {
-            holder.bind(numPadItems[position], { listener.onItemClick(numPadItems[position]) }, listener.isFingerPrintEnabled())
+            holder.bind(numPadItems[position], showFingerPrintButton) { listener.onItemClick(numPadItems[position]) }
         }
     }
 }
@@ -334,7 +359,7 @@ class NumPadItemViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
     private var imgFingerprint: ImageView = itemView.findViewById(R.id.imgFingerprint)
 
 
-    fun bind(item: NumPadItem, onClick: () -> (Unit), isFingerprintEnabled: Boolean) {
+    fun bind(item: NumPadItem, isFingerprintEnabled: Boolean, onClick: () -> (Unit)) {
 
         itemView.setOnClickListener { onClick.invoke() }
 
@@ -342,6 +367,7 @@ class NumPadItemViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
         txtLetters.visibility = View.GONE
         imgBackSpace.visibility = View.GONE
         imgFingerprint.visibility = View.GONE
+
         itemView.background = null
 
         when (item.type) {
