+++ app/src/main/java/bitcoin/wallet/modules/pin/PinInteractor.kt
@@ -1,14 +1,34 @@
 package bitcoin.wallet.modules.pin
 
-abstract class PinInteractor : PinModule.IInteractor {
+import bitcoin.wallet.core.IKeyStoreSafeExecute
+import bitcoin.wallet.core.IPinManager
 
-    var delegate: PinModule.IInteractorDelegate? = null
+class PinInteractor(
+        private val pinManager: IPinManager,
+        private val keystoreSafeExecute: IKeyStoreSafeExecute) : PinModule.IPinInteractor {
 
-    override fun viewDidLoad() {
+    var delegate: PinModule.IPinInteractorDelegate? = null
+    private var storedPin: String? = null
+
+    override fun set(pin: String?) {
+        storedPin = pin
     }
 
-    override fun onBackPressed() {
-       delegate?.onNavigateToPrevPage()
+    override fun validate(pin: String): Boolean {
+        return storedPin == pin
     }
 
+    override fun save(pin: String) {
+        keystoreSafeExecute.safeExecute(
+                action = Runnable {
+                    pinManager.store(pin)
+                },
+                onSuccess = Runnable { delegate?.didSavePin() },
+                onFailure = Runnable { delegate?.didFailToSavePin() }
+        )
+    }
+
+    override fun unlock(pin: String): Boolean {
+        return pinManager.validate(pin)
+    }
 }
