--- app/src/main/java/bitcoin/wallet/modules/pin/pinSubModules/EditPinPresenter.kt
@@ -1,13 +1,0 @@
-package bitcoin.wallet.modules.pin.pinSubModules
-
-import bitcoin.wallet.modules.pin.PinModule
-import bitcoin.wallet.modules.pin.PinPresenter
-
-class EditPinPresenter(interactor: PinModule.IInteractor, router: PinModule.IRouter) : PinPresenter(interactor, router) {
-
-    override fun updateViewTitleAndDescription() {
-        view?.setTitleForEditPin()
-        view?.setDescriptionForEditPin()
-    }
-
-}
