--- app/src/main/java/bitcoin/wallet/modules/pin/pinSubModules/UnlockPresenter.kt
@@ -1,19 +1,0 @@
-package bitcoin.wallet.modules.pin.pinSubModules
-
-import bitcoin.wallet.modules.pin.PinModule
-import bitcoin.wallet.modules.pin.PinPresenter
-
-class UnlockPresenter(interactor: PinModule.IInteractor, router: PinModule.IRouter) : PinPresenter(interactor, router) {
-
-    override fun viewDidLoad() {
-        super.viewDidLoad()
-
-        interactor.viewDidLoad()
-    }
-
-    override fun updateViewTitleAndDescription() {
-        view?.hideToolbar()
-        view?.setDescriptionForUnlock()
-    }
-
-}
