+++ app/src/main/java/bitcoin/wallet/modules/pin/set/SetPinPresenter.kt
@@ -1,9 +1,9 @@
-package bitcoin.wallet.modules.newpin.set
+package bitcoin.wallet.modules.pin.set
 
 import bitcoin.wallet.R
-import bitcoin.wallet.modules.newpin.ManagePinPresenter
-import bitcoin.wallet.modules.newpin.PinModule
-import bitcoin.wallet.modules.newpin.PinPage
+import bitcoin.wallet.modules.pin.ManagePinPresenter
+import bitcoin.wallet.modules.pin.PinModule
+import bitcoin.wallet.modules.pin.PinPage
 
 class SetPinPresenter(
         interactor: PinModule.IPinInteractor,
