+++ app/src/main/java/bitcoin/wallet/modules/pin/unlock/UnlockPinInteractor.kt
@@ -1,4 +1,4 @@
-package bitcoin.wallet.modules.newpin.unlock
+package bitcoin.wallet.modules.pin.unlock
 
 import bitcoin.wallet.core.*
