+++ app/src/test/java/bitcoin/wallet/modules/pin/PinInteractorTest.kt
@@ -1,4 +1,4 @@
-package bitcoin.wallet.modules.newpin
+package bitcoin.wallet.modules.pin
 
 import bitcoin.wallet.core.IKeyStoreSafeExecute
 import bitcoin.wallet.core.IPinManager
